requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
braceexpand==0.1.7
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
import-expression==1.1.4
Jinja2==3.1.2
jishaku==2.5.1
line-profiler==4.1.1
MarkupSafe==2.1.3
multidict==6.0.4
nodeenv==1.8.0
prisma==0.10.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
tomlkit==0.12.1
typing_extensions==4.8.0
yarl==1.9.2